python/pyproject.toml
[project]
name = "nexus"
version = "0.1.0"
description = "Nexus backend API - a reading and notes platform"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[postgresql-psycopg]>=2.0.36",
    "alembic>=1.14.0",
    "pydantic-settings>=2.6.0",
    "PyJWT[crypto]>=2.8.0",
    "httpx>=0.28.0",
    "structlog>=24.1.0",
    "lxml>=5.0.0",
    "pillow>=10.0.0",
    "pynacl>=1.5.0",
    "PyMuPDF>=1.24.0",
    "python-multipart>=0.0.22",
    "youtube-transcript-api>=1.2.4",
    "psycopg-binary>=3.1.0",
    "html5lib>=1.1",
    "stripe>=12.0.0",
    "llm-calling",
    "web-search-tool",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-xdist>=3.8.0",
    "httpx>=0.28.0",
    "pytest-httpserver>=1.0.0",
    "respx>=0.21.0",
    "pyright>=1.1.390",
    "pip-audit>=2.9.0",
    "zizmor>=1.24.1",
]

[project.scripts]
nexus = "nexus.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["nexus"]

[tool.uv.sources]
llm-calling = { git = "https://github.com/NielsdaWheelz/llm-calling", rev = "e61147978b9094f296911e8255e9aa596831ab1d" }
web-search-tool = { git = "https://github.com/NielsdaWheelz/web-search-tool", rev = "ce911344a2e31900eaee3039b5d61109ef54bcb1" }

[tool.ruff]
target-version = "py312"
line-length = 100
src = [".", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/conftest.py" = ["E402"]  # sys.path modification requires imports after

[tool.ruff.format]
quote-style = "double"

[tool.pyright]
typeCheckingMode = "standard"
pythonVersion = "3.12"
include = [
    "nexus/api",
    "nexus/db",
    "nexus/schemas",
    "nexus/tasks",
    "nexus/auth/extension.py",
    "nexus/auth/permissions.py",
    "nexus/auth/verifier.py",
    "nexus/cli.py",
    "nexus/errors.py",
    "nexus/jobs/registry.py",
    "nexus/jobs/worker.py",
    "nexus/logging.py",
    "nexus/middleware/request_id.py",
    "nexus/responses.py",
    "nexus/services/api_key_resolver.py",
    "nexus/services/capabilities.py",
    "nexus/services/canonicalize.py",
    "nexus/services/chat_runs.py",
    "nexus/services/conversation_branches.py",
    "nexus/services/chat_prompt.py",
    "nexus/services/context_rendering.py",
    "nexus/services/context_window.py",
    "nexus/services/contexts.py",
    "nexus/services/content_indexing.py",
    "nexus/services/contributor_credits.py",
    "nexus/services/contributors.py",
    "nexus/services/crypto.py",
    "nexus/services/epub_ingest.py",
    "nexus/services/epub_lifecycle.py",
    "nexus/services/extension_sessions.py",
    "nexus/services/fragment_blocks.py",
    "nexus/services/image_proxy.py",
    "nexus/services/ingest_recovery.py",
    "nexus/services/metadata_enrichment.py",
    "nexus/services/message_context_items.py",
    "nexus/services/node_ingest.py",
    "nexus/services/notes.py",
    "nexus/services/object_links.py",
    "nexus/services/object_refs.py",
    "nexus/services/oracle.py",
    "nexus/services/pdf_highlight_geometry.py",
    "nexus/services/pdf_lifecycle.py",
    "nexus/services/pdf_locking.py",
    "nexus/services/pdf_quote_match.py",
    "nexus/services/pdf_quote_match_policy.py",
    "nexus/services/pdf_readiness.py",
    "nexus/services/playback_source.py",
    "nexus/services/quote_context_errors.py",
    "nexus/services/reader.py",
    "nexus/services/redact.py",
    "nexus/services/rss_transcript_fetch.py",
    "nexus/services/semantic_chunks.py",
    "nexus/services/seq.py",
    "nexus/services/shares.py",
    "nexus/services/locator_resolver.py",
    "nexus/services/transcript_media.py",
    "nexus/services/transcript_segments.py",
    "nexus/services/url_normalize.py",
    "nexus/services/user_keys.py",
    "nexus/services/users.py",
    "nexus/services/x_identity.py",
    "nexus/services/youtube_identity.py",
    "nexus/services/youtube_transcripts.py",
    "nexus/storage/paths.py",
]
exclude = [
    "tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short --strict-markers -m \"not supabase and not network\""
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "unit: pure logic tests, no DB/network",
    "integration: DB/API-backed tests",
    "slow: tests that are materially slower than normal local feedback loops",
    "supabase: requires Supabase local auth/storage services",
    "network: requires internet access (hits real external services)",
    "real_media: strict acceptance tests over real media artifacts and evidence traces",
    "live_provider: strict acceptance tests over live external providers",
]
